+++ Hippo.Web/ClientApp/src/components/Order/OrderForm/OrderFormField.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { useFormContext } from "react-hook-form";
-import FormField from "../../../Shared/Form/FormField";
-import { FormFieldProps } from "../../../Shared/Form/formTypes";
+import HipFormField from "../../../Shared/Form/HipFormField";
+import { HipFormFieldProps } from "../../../Shared/Form/formTypes";
 import { OrderModel } from "../../../types";
 
-type OrderFormFieldProps = FormFieldProps<OrderModel> & {};
+type OrderFormFieldProps = HipFormFieldProps<OrderModel> & {};
 
 const OrderFormField: React.FC<OrderFormFieldProps> = ({ ...props }) => {
   const {
@@ -13,7 +13,7 @@ const OrderFormField: React.FC<OrderFormFieldProps> = ({ ...props }) => {
   } = useFormContext<OrderModel>();
 
   return (
-    <FormField register={register} error={errors[props.name]} {...props} />
+    <HipFormField register={register} error={errors[props.name]} {...props} />
   );
 };
